Decouple asset loaders (spv, obj and image)
